+++ spirit/user/auth/forms.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 
 from django import forms
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+from django.contrib.auth.forms import AuthenticationForm
 from django.utils.translation import ugettext_lazy as _
 from django.conf import settings
 
@@ -13,8 +13,16 @@ from ..forms import CleanEmailMixin
 User = get_user_model()
 
 
-class RegistrationForm(CleanEmailMixin, UserCreationForm):
+class RegistrationForm(CleanEmailMixin, forms.ModelForm):
 
+    email2 = forms.CharField(
+        label=_("Email confirmation"),
+        widget=forms.EmailInput,
+        max_length=254,
+        help_text=_("Enter the same email as above, for verification.")
+    )
+    # todo: add password validator for Django 1.9
+    password = forms.CharField(label=_("Password"), widget=forms.PasswordInput)
     honeypot = forms.CharField(label=_("Leave blank"), required=False)
 
     class Meta:
@@ -23,7 +31,7 @@ class RegistrationForm(CleanEmailMixin, UserCreationForm):
 
     def __init__(self, *args, **kwargs):
         super(RegistrationForm, self).__init__(*args, **kwargs)
-        self.fields['email'].required = True  # Django model does not required this by default
+        self.fields['email'].required = True  # Django model does not requires it
 
     def clean_honeypot(self):
         """Check that nothing has been entered into the honeypot."""
@@ -46,8 +54,23 @@ class RegistrationForm(CleanEmailMixin, UserCreationForm):
 
         return username
 
+    def clean_email2(self):
+        email = self.cleaned_data.get("email")
+        email2 = self.cleaned_data["email2"]
+
+        if settings.ST_CASE_INSENSITIVE_EMAILS:
+            email2 = email2.lower()
+
+        if email and email != email2:
+            raise forms.ValidationError(
+                _("The two email fields didn't match.")
+            )
+
+        return email2
+
     def save(self, commit=True):
         self.instance.is_active = False
+        self.instance.set_password(self.cleaned_data["password"])
         return super(RegistrationForm, self).save(commit)